refactor : 패키지 구체화
