+++ src/main/java/model/Deck/Deck.java
@@ -1,4 +1,4 @@
-package model;
+package model.Deck;
 
 import java.util.ArrayList;
 import java.util.Collections;
