+++ src/main/java/model/participant/Participant.java
@@ -1,8 +1,10 @@
-package model;
+package model.participant;
 
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.IntStream;
+import model.Deck.Card;
+import model.Deck.Deck;
 
 public class Participant {
     private static final int INITIAL_DEAL_CARD_COUNT = 2;
