+++ src/main/java/model/participant/ParticipantHand.java
@@ -1,8 +1,10 @@
-package model;
+package model.participant;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import model.Deck.Card;
+import model.Deck.CardRank;
 
 public final class ParticipantHand {
     private static final int BURST_SCORE_LIMIT = 21;
