+++ src/test/java/model/DealerTest.java
@@ -2,7 +2,10 @@ package model;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-import org.junit.jupiter.api.BeforeAll;
+import model.Deck.Card;
+import model.Deck.CardRank;
+import model.Deck.CardSuit;
+import model.participant.Dealer;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
