+++ src/test/java/model/DeckTest.java
@@ -5,6 +5,10 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import model.Deck.Card;
+import model.Deck.CardRank;
+import model.Deck.CardSuit;
+import model.Deck.Deck;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
