+++ src/test/java/model/ParticipantHandTest.java
@@ -2,6 +2,10 @@ package model;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import model.Deck.Card;
+import model.Deck.CardRank;
+import model.Deck.CardSuit;
+import model.participant.ParticipantHand;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
