+++ src/test/java/model/ParticipantWinningResultTest.java
@@ -4,6 +4,14 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
 import java.util.Map;
+import model.Deck.Card;
+import model.Deck.CardRank;
+import model.Deck.CardSuit;
+import model.participant.Dealer;
+import model.participant.Player;
+import model.participant.Players;
+import model.result.GameResult;
+import model.result.ParticipantWinningResult;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
